小学期项目/finally! - 部署版/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
jieba>=0.42.1
python-dotenv>=0.19.0
openai>=1.0.0
langchain-openai>=0.0.1
pymysql>=1.0.2
sqlalchemy>=2.0.0
configparser>=5.0.0
requests>=2.28.1
charset-normalizer>=2.0.4
cryptography>=39.0.1
idna>=3.4
setuptools>=65.6.3
urllib3>=1.26.15
wheel>=0.38.4
